+++ lib/twistedconjugacy.gd
@@ -69,9 +69,9 @@ DeclareOperation(
     "CoincidenceReidemeisterSpectrumOp",
     [ IsGroup, IsGroup ]
 );
-DeclareGlobalFunction( "FullReidemeisterSpectrum" );
+DeclareGlobalFunction( "TotalReidemeisterSpectrum" );
 DeclareOperation(
-    "FullReidemeisterSpectrumOp",
+    "TotalReidemeisterSpectrumOp",
     [ IsGroup ]
 );
 
@@ -125,13 +125,7 @@ DeclareOperation(
 
 DeclareGlobalFunction( "InducedHomomorphism" );
 DeclareGlobalFunction( "RestrictedHomomorphism" );
+DeclareGlobalFunction( "InclusionHomomorphism" );
 
 
-DeclareProperty( "IsNilpotentByFinite", IsGroup );
-InstallTrueMethod( IsNilpotentByFinite, IsNilpotentGroup );
-InstallTrueMethod( IsNilpotentByFinite, IsGroup and IsFinite );
-DeclareProperty( "IsPolycyclicByFinite", IsGroup );
-InstallTrueMethod( IsPolycyclicByFinite, IsPolycyclicGroup );
-InstallTrueMethod( IsPolycyclicByFinite, IsGroup and IsFinite );
 DeclareProperty( "IsNilpotentByAbelian", IsGroup );
-InstallTrueMethod( IsNilpotentByAbelian, IsNilpotentGroup );