+++ CHANGES.md
@@ -1,6 +1,24 @@
 This file describes changes in the GAP package TwistedConjugacy.
 
 
+2.4.0 (2024-12-05)
+------------------
+
+- Added methods for intersections of arbitrary subgroups and cosets in
+  PcpGroups
+- Added method to test membership in a double coset of a PcpGroup
+
+- Algorithms for infinite polycyclic groups again require the Polycyclic
+  package, since many methods would otherwise run forever due to the lack of
+  finiteness checks in existing methods
+
+- Renamed FullReidemeisterSpectrum to TotalReidemeisterSpectrum, to avoid
+  confusion with a spectrum being 'full'
+  
+- Replaced ToggleSafeMode@TwistedConjugacy by variable ASSERT@TwistedConjugacy
+
+
+
 2.3.0 (2024-09-17)
 ------------------
 
@@ -12,6 +30,7 @@ This file describes changes in the GAP package TwistedConjugacy.
 
 - Added ToggleSafeMode@TwistedConjugacy, to use built-in assertions (including
   those in Polycyclic) to check for any errors.
+
 
 
 2.2.0 (2024-02-13)
@@ -29,6 +48,7 @@ This file describes changes in the GAP package TwistedConjugacy.
 - Added IsNilpotentByFinite and IsPolycyclicByFinite properties
 - Code for infinite groups is now less dependent on the Polycyclic package
   being loaded
+
 
 
 2.1.0 (2023-03-16)
@@ -53,6 +73,7 @@ This file describes changes in the GAP package TwistedConjugacy.
 - CoincidenceGroup can now take 3 or more endomorphisms as argument
 - RepresentativeTwistedConjugation and IsTwistedConjugate can now take lists
   as input, and will try to solve the multiple twisted conjugacy problem
+
 
 
 2.0.0 (2021-05-01)
@@ -110,6 +131,7 @@ This file describes changes in the GAP package TwistedConjugacy.
 - Finite PcpGroups will now automatically be converted to PcGroups
 
 
+
 1.0.1 (2020-06-07)
 ------------------
 
@@ -117,6 +139,7 @@ This file describes changes in the GAP package TwistedConjugacy.
 - Test Suite now has full coverage
 - Some functions in HelpFunctions.gi now have local scope
 - Updated documentation to include examples
+
 
 
 1.0.0 (2020-02-19)